refactory stage 1 - not working version
