+++ server/backend/models/project.js
@@ -3,9 +3,11 @@ const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
 const ProjectSchema = new mongoose.Schema({
-  name: String,
-  tenantid: Schema.Types.ObjectId,
-  type: { type: Number, default: 1 },
+  name: { type: String, required: true },
+  tenantId: { type: Schema.Types.ObjectId, required: true },
+  projectType: { type: Number, required: true },
+  createdAt: { type: Date, required: true },
+  udpatedAt: { type: Date, required: true },
 });
 
 module.exports = mongoose.model('project', ProjectSchema);
