+++ server/backend/models/resource.js
@@ -3,7 +3,7 @@ const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
 const ResourceSchema = new mongoose.Schema({
-  projectid: { type: Schema.Types.ObjectId, required: true },
+  projectId: { type: Schema.Types.ObjectId, required: true, unique: true },
   name: { type: String, required: true },
   resourceType: { type: Number, required: true }, // 1 - person, no definition for other contacts
   contacts: [{
@@ -11,6 +11,8 @@ const ResourceSchema = new mongoose.Schema({
     value: { type: String, required: true },
   }, // 1 - email, no definition for other contacts
   ],
+  createdAt: { type: Date, required: true },
+  udpatedAt: { type: Date, required: true },
 });
 
 module.exports = mongoose.model('resource', ResourceSchema);
