+++ server/backend/models/tenant.js
@@ -1,8 +1,10 @@
 const mongoose = require('mongoose');
 
 const TenantSchema = new mongoose.Schema({
-  name: String,
-  type: { type: Number, default: 1 },
+  name: { type: String, required: true },
+  tenantType: { type: Number, required: true },
+  createdAt: { type: Date, required: true },
+  udpatedAt: { type: Date, required: true },
 });
 
 module.exports = mongoose.model('tenant', TenantSchema);
