+++ server/backend/utils/handleError.js
@@ -1,0 +1,6 @@
+const logger = require('../../logger');
+
+module.exports = (response, error) => {
+  logger.error(error);
+  response.status(500).send({ err: error });
+};
